api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyodbc>=4.0.39
pyyaml>=6.0
